refactor(api): move extmark specific functions to api/extmark.c
